Added user authentication to getById, delete and update methods

+++ src/tasks/tasks.service.ts
@@ -22,8 +22,8 @@ export class TasksService {
         return this.taskRepository.getTasks(filterDto, user)
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id}})
 
         if(!found) {
             throw new NotFoundException(`Task with id ${id} not found`)
@@ -36,8 +36,8 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async deleteTask(id: number): Promise<Task[]> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTask(id: number, user: User): Promise<Task[]> {
+        const result = await this.taskRepository.delete({ id, userId: user.id })
         
         if(result.affected === 0) {
             throw new NotFoundException(`Task with id ${id} doesn't exist`)
@@ -47,8 +47,12 @@ export class TasksService {
         
      }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id)
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user)
         task.status = status
 
         await task.save()
